syncing usage example and minor cleanup

+++ groups/btl/btlsos/btlsos_tcpcbacceptor.cpp
@@ -481,7 +481,7 @@ void TcpCbAcceptor::cancelAll()
                           d_callbacks.begin() + d_callbacks.size() - 1);
         BSLS_ASSERT(d_currentRequest_p == d_callbacks.back());
 
-        int numToCancel = toBeCancelled.size();
+        int numToCancel = static_cast<int>(toBeCancelled.size());
         while (--numToCancel >= 0) {
             TcpCbAcceptor_Reg *reg = toBeCancelled[numToCancel];
             BSLS_ASSERT(reg);
@@ -494,7 +494,7 @@ void TcpCbAcceptor::cancelAll()
         bsl::deque<TcpCbAcceptor_Reg *>
                                      toBeCancelled(d_callbacks, d_allocator_p);
         d_callbacks.clear();
-        int numToCancel = toBeCancelled.size();
+        int numToCancel = static_cast<int>(toBeCancelled.size());
         if (numToCancel) {
             d_manager_p->deregisterSocketEvent(d_serverSocket_p->handle(),
                                                btlso::EventType::BTESO_ACCEPT);
